FailedDeliveriesScreen is now AlertCenterScreen
