+++ lib/shared_components/constants/ui_strings.dart
@@ -80,4 +80,4 @@ const kLogOutAlertDialog =
 const kNoDescription = 'No description';
 const kSomethingWentWrong = 'Something went wrong';
 const kFailedDeliveriesNote =
-    'Sometimes when AnonAddy attempts to send an email, the delivery is not successful.\nThis is often referred to as a "bounced email".';
+    'Sometimes when AnonAddy attempts to send an email, the delivery is not successful. This is often referred to as a "bounced email".';
